#16 MyHomePageをHomeScreenにリネームして、別クラスに切り分けた
